refactor: feature/admin/index.ts 수정 및 적용
