+++ frontend/src/features/admin/index.ts
@@ -6,3 +6,9 @@ export { default as AdminSidebar } from './components/AdminSidebar';
 export { default as AdminTable } from './components/AdminTable';
 export { default as AdminReportTableBody } from './components/AdminReportTableBody';
 export { default as AdminUnverifiedReviewsTableBody } from './components/AdminUnverifiedReviewsTableBody';
+
+export * from './api/admin.api';
+
+export * from './hooks/useAdmin';
+
+export * from './model/admin.model';
